[iOS] Don't deadlock trying to swap tracks
If you repeatedly swapped tracks it looks like you could
cause a deadlock and/or lengthy hang.

When swapping tracks we'd dispose the old player before
spinning up a new one. This disposing happened on the
main thread so we could have sane semantics about ownership
throughout the pretty heavily threaded AudioPlayer class.

It looks like `HandleBufferCompleted` was being invoked either
as part of invoking `outputQueue.Stop (true)`, or maybe it was
invoked just before we tried disposing the player. Either way,
the result was that if we hit the `Buffering = true` line we
would try to emit an event on the main thread using our
SyncContext, but the main thread was blocked because
`outputQueue.Stop` would only return once the `HandleBufferCompleted`
event returned.

The fix is simply to make the buffering event emit asynchronously
instead of synchronously. That was the original intention anyway.
It looks like it was just an unfortunate accident that it used
`Send` instead of `Post`. Those are two very very badly named methods